+++ src/lib/Services/Object/PasswordRevisionService.php
@@ -64,7 +64,7 @@ class PasswordRevisionService extends AbstractRevisionService {
      * @param int    $edited
      * @param bool   $hidden
      * @param bool   $trashed
-     * @param bool   $favourite
+     * @param bool   $favorite
      *
      * @return PasswordRevision
      *
@@ -84,12 +84,12 @@ class PasswordRevisionService extends AbstractRevisionService {
         int $edited,
         bool $hidden,
         bool $trashed,
-        bool $favourite
+        bool $favorite
     ): PasswordRevision {
         if($cseType === EncryptionService::CSE_ENCRYPTION_NONE) $hash = sha1($password);
 
         $revision = $this->createModel(
-            $model, $password, $username, $cseType, $hash, $label, $url, $notes, $customFields, $folder, $edited, $hidden, $trashed, $favourite
+            $model, $password, $username, $cseType, $hash, $label, $url, $notes, $customFields, $folder, $edited, $hidden, $trashed, $favorite
         );
 
         $revision = $this->validationService->validatePassword($revision);
@@ -112,7 +112,7 @@ class PasswordRevisionService extends AbstractRevisionService {
      * @param int    $edited
      * @param bool   $hidden
      * @param bool   $trashed
-     * @param bool   $favourite
+     * @param bool   $favorite
      *
      * @return PasswordRevision
      */
@@ -130,7 +130,7 @@ class PasswordRevisionService extends AbstractRevisionService {
         int $edited,
         bool $hidden,
         bool $trashed,
-        bool $favourite
+        bool $favorite
     ): PasswordRevision {
 
         $revision = new PasswordRevision();
@@ -154,7 +154,7 @@ class PasswordRevisionService extends AbstractRevisionService {
         $revision->setNotes($notes);
         $revision->setCustomFields($customFields);
         $revision->setFolder($folder);
-        $revision->setFavourite($favourite);
+        $revision->setFavorite($favorite);
         $revision->setClient('');
         $revision->setEdited($edited);
         $revision->setSseType(EncryptionService::DEFAULT_SSE_ENCRYPTION);