+++ src/vue/App.vue
@@ -11,8 +11,8 @@
                 <router-link class="nav-icon-recent" :to="{ name: 'Recent'}" active-class="active" tag="li">
                     <translate say="Recent"/>
                 </router-link>
-                <router-link class="nav-icon-favourites" :to="{ name: 'Favourites'}" active-class="active" tag="li">
-                    <translate say="Favourites"/>
+                <router-link class="nav-icon-favorites" :to="{ name: 'Favorites'}" active-class="active" tag="li">
+                    <translate say="Favorites"/>
                 </router-link>
                 <router-link class="nav-icon-shared" :to="{ name: 'Shared'}" active-class="active" tag="li">
                     <translate say="Shared"/>
@@ -23,7 +23,7 @@
                 <router-link class="nav-icon-security" :to="{ name: 'Security'}" active-class="active" tag="li">
                     <translate say="Security"/>
                 </router-link>
-                <router-link class="nav-icon-search" :to="{ name: 'Search'}" active-class="active" tag="li" v-if="showSearch">
+                <router-link class="nav-icon-search" :to="{ name: 'Search'}" active-class="active" tag="li" v-if="isSearchVisible">
                     <translate say="Search"/>
                 </router-link>
             </ul>
@@ -68,17 +68,23 @@
         },
 
         data() {
-            let serverVersion = SettingsManager.get('server.version');
+            let serverVersion = SettingsManager.get('server.version'),
+                showSearch    = SettingsManager.get('client.search.show');
 
             return {
-                serverVersion: serverVersion,
-                showMore     : false
+                serverVersion,
+                showSearch,
+                showMore: false
             };
         },
 
+        created() {
+            SettingsManager.observe('client.search.show', (v) => { this.showSearch = v.value; });
+        },
+
         computed: {
-            showSearch() {
-                return this.$route.name === 'Search';
+            isSearchVisible() {
+                return this.$route.name === 'Search' || this.showSearch;
             }
         },
 
@@ -89,9 +95,6 @@
                 } else {
                     Utility.openLink('https://github.com/marius-wieschollek/passwords-webextension/wiki/chromium-builds');
                 }
-            },
-            openWikiPage() {
-                Utility.openLink('https://git.mdns.eu/nextcloud/passwords/wikis/home#users');
             }
         }
     };
@@ -127,7 +130,7 @@
             &.nav-icon-tags:before { content : "\f02c"; }
             &.nav-icon-security:before { content : "\f132"; }
             &.nav-icon-shared:before { content : "\f1e0"; }
-            &.nav-icon-favourites:before { content : "\f005"; }
+            &.nav-icon-favorites:before { content : "\f005"; }
             &.nav-icon-search:before { content : "\f002"; }
             &.nav-icon-trash:before { content : "\f014"; }
             &.nav-icon-more:before { content : "\f067"; }